Fixed ConfirmRegistration and DisplayMessageActivity
I fixed M5 and got the ConfirmRegistration screen to work properly by
copying over existing code from the DisplayMessageActivity class and
editing it as needed. By doing that I was able to use the
ConfirmRegistration screen as the landing page after they registered
and the DisplayMessageActivity screen as the landing page after they
login. With this commit M5 should be complete